feat: also gate participant.ts not only tests
